Ensure bookmarks are updated before processing post-suspension-tasks
Also updated unit test to make sure the workflow registry mock returns the workflow blueprint.

TODO: Figure out a way to wait until Rebus processed the dispatched `ExecuteWorkflowInstanceRequest` message (which happens asynchronously in-memory). Until then, yielding control here and waiting for a few seconds works too.
